pyproject.toml
[project]
name = "auto-contributor"
version = "0.1.0"
description = "Automated GitHub contribution bot using Claude Code CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "apscheduler",
    "structlog",
    "rich",
    "typer",
    "gitpython",
]

[project.scripts]
auto-contributor = "auto_contributor.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_contributor"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "ASYNC", # flake8-async
]

[tool.ruff.lint.isort]
known-first-party = ["auto_contributor"]

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
